+++ ChangeLog.md
@@ -1,4 +1,8 @@
-### 2016-05-18
+### 2016-05-19 v1.1
+	* Added example with starting point of {0.0, 0.0}
+	* Renamed test example to rosenbrock
+
+### 2016-05-18 v1.0
 	* Re-evaluate all vertices after an unsuccessful contraction.
 	* Initial release to github, v1.0